pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "storytelling-app"
version = "0.1.0"
description = "A collaborative AI-driven storytelling application"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic-ai",
    "python-dotenv",
    "click",
    "httpx",
    "logfire",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[project.scripts]
storytelling = "src.main:main"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true